Better level generation + Comments
