+++ desktop/src/com/moreoptions/prototype/desktop/DesktopLauncher.java
@@ -3,10 +3,12 @@ package com.moreoptions.prototype.desktop;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import com.moreoptions.prototype.MoreOptions;
+import com.moreoptions.prototype.level.Map;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
-		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		new LwjglApplication(new MoreOptions(), config);
+		new Map();
+		// LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+		// new LwjglApplication(new MoreOptions(), config);
 	}
 }
